appearance_checker_backend/requirements.txt
Flask==1.1.2
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-language==2.0.0